+++ example/lib/main.dart
@@ -41,7 +41,9 @@ class _MyHomePageState extends State<MyHomePage> {
           // in the middle of the parent.
           child: Column(
             children: [
-              DotLottieLoader.fromAsset("assets/anim2.lottie",
+
+              ///Simple .Lottie Animation
+              DotLottieLoader.fromAsset("assets/animation.lottie",
                   frameBuilder: (ctx, dotlottie) {
                 if (dotlottie != null) {
                   return Lottie.memory(dotlottie.animations.values.single);
@@ -50,20 +52,52 @@ class _MyHomePageState extends State<MyHomePage> {
                 }
               }),
               const SizedBox(height: 20),
+
+              ///.Lottie Animation From network
               DotLottieLoader.fromNetwork(
                 "https://github.com/sartajroshan/dotlottieloader-flutter/raw/master/example/assets/animation.lottie",
                 frameBuilder: (ctx, dotlottie) {
                   if (dotlottie != null) {
-                    return Lottie.memory(dotlottie!.animations.values.single);
+                    return Lottie.memory(dotlottie!.animations.values.single,
+                    );
                   } else {
                     return Container();
                   }
                 },
                 errorBuilder: (ctx, e, s) {
-                  print(s);
                   return Text(e.toString());
                 },
               ),
+              const SizedBox(height: 20),
+
+              ///.lottie animation with external image
+              DotLottieLoader.fromAsset("assets/animation_external_image.lottie",
+                  frameBuilder: (ctx, dotlottie) {
+                    if (dotlottie != null) {
+                      return Lottie.memory(dotlottie.animations.values.single,
+                          imageProviderFactory: (asset) {
+                            return MemoryImage(dotlottie.images[asset.fileName]!);
+                          }
+                      );
+                    } else {
+                      return Container();
+                    }
+                  }),
+              const SizedBox(height: 20),
+
+              ///.lottie animation with inline image
+              DotLottieLoader.fromAsset("assets/animation_inline_image.lottie",
+                  frameBuilder: (ctx, dotlottie) {
+                    if (dotlottie != null) {
+                      return Lottie.memory(dotlottie.animations.values.single,
+                        imageProviderFactory: (asset) {
+                       return MemoryImage(dotlottie.images[asset.fileName]!);
+                        }
+                      );
+                    } else {
+                      return Container();
+                    }
+                  }),
             ],
           ),
         ),