+++ lib/dotlottie_loader.dart
@@ -1,2 +1,4 @@
-export 'models/dotlottie_models.dart' show DotLottie;
-export 'dotlottie_loader_builder.dart' show DotLottieLoader;
+library dotlottie_loader;
+
+export 'src/models/dotlottie_models.dart' show DotLottie;
+export 'src/dotlottie_loader_builder.dart' show DotLottieLoader;