+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Added examples with inline and external images.
+* File structure change
+* repository link corrected
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* initial release